chore(frontend): make vue-tsc a real type-check (config + plugin split)
The frontend type-check was a no-op: the references-root tsconfig has no
`include`, so `vue-tsc --noEmit` checked 0 files and always passed. Nothing
ever type-checked the frontend — the build is `vite build` (esbuild strips
types), no CI workflow runs vue-tsc, and the local hook ritual ran the no-op
form. Latent type errors accumulated invisibly (this is how the updateRecipeImage
break shipped).

Make `vue-tsc --build --noEmit` actually run:
- ignoreDeprecations "6.0" (TS 6.0.3) silences the fatal baseUrl TS5101 in
  tsconfig.app/vitest.
- Split src/types/Plugins.ts into pure types + src/plugin_registry.ts (runtime
  import.meta.glob + TANDOOR_PLUGINS), so vite.config.ts (node project) imports
  only types — fixes TS6307/TS2339 at the project boundary. Registry lives at
  src/plugin_registry.ts, not src/plugins/ (gitignored plugin mount point).
  Repointed 5 consumers.
- Add src/types/Plugins.ts to tsconfig.node include.
- .gitignore safety-net for vue-tsc --build emit.

No behavior change. Kept as a per-file verification tool; the ~1000 pre-existing
errors (mostly upstream-origin latent debt) are intentionally not addressed here.
