+++ vue3/src/types/Plugins.ts
@@ -25,11 +25,3 @@ export type TandoorPlugin = {
 export type PluginModule = {
   plugin: TandoorPlugin
 }
-
-const pluginModules = import.meta.glob('@/plugins/*/plugin.ts', { eager: true })
-export let TANDOOR_PLUGINS = [] as TandoorPlugin[]
-Object.values(pluginModules).forEach(module => {
-    if(!module.plugin.disabled){
-        TANDOOR_PLUGINS.push(module.plugin)
-    }
-})
