+++ vue3/tsconfig.app.json
@@ -5,6 +5,7 @@
   "compilerOptions": {
     "composite": true,
     "verbatimModuleSyntax": false,
+    "ignoreDeprecations": "6.0",
     "baseUrl": ".",
     "paths": {
       "@/*": ["./src/*"]
