+++ vue3/tsconfig.vitest.json
@@ -3,6 +3,7 @@
   "exclude": [],
   "compilerOptions": {
     "composite": true,
+    "ignoreDeprecations": "6.0",
     "lib": [],
     "types": ["node", "jsdom", "vitest/globals"]
   }
